+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+<a name="2.1.3"></a>
+## [2.1.3](https://github.com/adonisjs/adonis-commands/compare/v2.1.2...v2.1.3) (2016-12-12)
+
+
+### Bug Fixes
+
+* **make:model:** stay consistent with file names ([54a6c8f](https://github.com/adonisjs/adonis-commands/commit/54a6c8f)), closes [#9](https://github.com/adonisjs/adonis-commands/issues/9)
+
+
+### Features
+
+* **websocket:** add support for websocket controller ([b5e92e2](https://github.com/adonisjs/adonis-commands/commit/b5e92e2))
+
+
+
+<a name="2.1.2"></a>
+## [2.1.2](https://github.com/adonisjs/adonis-commands/compare/v2.1.1...v2.1.2) (2016-09-26)
+
+
+
+<a name="2.1.1"></a>
+## [2.1.1](https://github.com/adonisjs/adonis-commands/compare/v2.0.5...v2.1.1) (2016-09-26)
+
+
+### Features
+
+* add support for nested scaffolds ([622f349](https://github.com/adonisjs/adonis-commands/commit/622f349))
+* **commands:** add route:list command ([#8](https://github.com/adonisjs/adonis-commands/issues/8)) ([91ffdd0](https://github.com/adonisjs/adonis-commands/commit/91ffdd0))
+
+
 <a name="2.1.0"></a>
 # [2.1.0](https://github.com/adonisjs/adonis-commands/compare/v2.0.5...v2.1.0) (2016-06-26)